requirements.txt
torch==2.3.1+cu118
torchaudio==2.3.1+cu118
torchvision==0.18.1+cu118
transformers==4.47.1
pandas==2.2.3
numpy==1.26.3
matplotlib==3.10.0
seaborn==1.3.1
tqdm==4.67.1
pillow==11.1.0
scikit-learn==1.6.0